+++ src/fourchess/pieces/Queen.groovy
@@ -18,16 +18,16 @@ public class Queen extends Piece
 	public Set<Point> getValidMoves(Point position)
 	{
 		Set<Point> moves = super.getValidMoves(position);
-		Piece piece = board.getPiece(position.x, position.y);
+		Piece piece = board.getPiece(position.x.intValue(), position.y.intValue());
 		
-		Rules.checkPath(piece, position.x, position.y, 0, -1, moves);  // Up
-		Rules.checkPath(piece, position.x, position.y, 0, 1, moves);   // Down
-		Rules.checkPath(piece, position.x, position.y, -1, 0, moves);  // Left
-		Rules.checkPath(piece, position.x, position.y, 1, 0, moves);   // Right
-		Rules.checkPath(piece, position.x, position.y, -1, -1, moves); // Up-Left
-		Rules.checkPath(piece, position.x, position.y, -1, 1, moves);  // Down-Left
-		Rules.checkPath(piece, position.x, position.y, 1, -1, moves);  // Up-Right 
-		Rules.checkPath(piece, position.x, position.y, 1, 1, moves);   // Down-Right
+		Rules.checkPath(piece, position.x.intValue(), position.y.intValue(), 0, -1, moves);  // Up
+		Rules.checkPath(piece, position.x.intValue(), position.y.intValue(), 0, 1, moves);   // Down
+		Rules.checkPath(piece, position.x.intValue(), position.y.intValue(), -1, 0, moves);  // Left
+		Rules.checkPath(piece, position.x.intValue(), position.y.intValue(), 1, 0, moves);   // Right
+		Rules.checkPath(piece, position.x.intValue(), position.y.intValue(), -1, -1, moves); // Up-Left
+		Rules.checkPath(piece, position.x.intValue(), position.y.intValue(), -1, 1, moves);  // Down-Left
+		Rules.checkPath(piece, position.x.intValue(), position.y.intValue(), 1, -1, moves);  // Up-Right
+		Rules.checkPath(piece, position.x.intValue(), position.y.intValue(), 1, 1, moves);   // Down-Right
 		
 		return moves;
 	}
