+++ src/fourchess/pieces/Rook.groovy
@@ -18,12 +18,12 @@ public class Rook extends Piece
 	public Set<Point> getValidMoves(Point position)
 	{
 		Set<Point> moves = super.getValidMoves(position);
-		Piece piece = board.getPiece(position.x, position.y);
+		Piece piece = board.getPiece(position.x.intValue(), position.y.intValue());
 		
-		Rules.checkPath(piece, position.x, position.y, 0, -1, moves); // Up
-		Rules.checkPath(piece, position.x, position.y, 0, 1, moves);  // Down
-		Rules.checkPath(piece, position.x, position.y, -1, 0, moves); // Left
-		Rules.checkPath(piece, position.x, position.y, 1, 0, moves);  // Right
+		Rules.checkPath(piece, position.x.intValue(), position.y.intValue(), 0, -1, moves); // Up
+		Rules.checkPath(piece, position.x.intValue(), position.y.intValue(), 0, 1, moves);  // Down
+		Rules.checkPath(piece, position.x.intValue(), position.y.intValue(), -1, 0, moves); // Left
+		Rules.checkPath(piece, position.x.intValue(), position.y.intValue(), 1, 0, moves);  // Right
 		
 		return moves;
 	}
